// types/forms.ts
export type ShippingFormData = {
  _id?: string;
  name: string;
  slug: string;
  description: string;
  image: string;
  status: "draft" | "publish" | "archive";
  user_id?: string | null | undefined;
  delay: number;
  fixed_amount: number;
  unit_price_weight: number;
  price_range_start: number;
  price_range_end: number;
  fees: number;
  region: string[];
  createdAt?: string;
  store?: string;
};