// types/schemas.ts
export const shippingValidationSchema: ZodType<ShippingFormData> = z.object({
  name: z.string().regex(nameFormat, {
    message: "format: no special (&/,), min 4 and max 60 characters",
  }),
  slug: z.string().regex(slugFormat, {
    message:
      "format: no blank space (_), no special character (&,',/|...), min 4 and max 20 characters",
  }),
  description: z.string().regex(descriptionFormat, {
    message: "format:min 4 and max 500 characters",
  }),
  image: z.string().url("Image url is invalid"),
  status: z.enum(["draft", "publish", "archive"]),
  user_id: z.string().optional().nullable(),
  delay: z.coerce.number().min(0).max(360, "less than 360 days"),
  fees: z.coerce.number().min(0).max(100000),
  region: z.string().array(),
  fixed_amount: z.coerce.number().min(0).max(100000),
  unit_price_weight: z.coerce.number().min(0).max(100000),
  price_range_start: z.coerce.number().min(0).max(100000),
  price_range_end: z.coerce.number().min(0).max(100000),
  store: z.string().optional(),
});